feat: Add local authentication and fix OVNCP to run fully
- Added local authentication with admin/admin credentials
- Fixed all Go compilation errors in middleware and handlers
- Fixed TypeScript errors in web UI components
- Made OAuth providers optional, allowing local auth only
- Implemented auto-migration on API startup
- Added local login UI with username/password form
- Fixed route parameter conflicts in API
- Created missing middleware functions (auth, request_id, error_handler)
- Fixed API client TypeScript types
- Updated docker-compose with proper auth configuration

The OVNCP stack now runs fully with:
- API server with JWT authentication
- React web UI with local login support
- PostgreSQL with automatic migrations
- Multi-tenancy and RBAC support

+++ internal/middleware/audit.go
@@ -302,7 +302,7 @@ func Audit(cfg AuditConfig) gin.HandlerFunc {
 		parseActionAndResource(c, event)
 		
 		// Add request body (redact sensitive fields)
-		if cfg.LogRequestBody && len(requestBody) > 0 && len(requestBody) <= cfg.MaxBodySize {
+		if cfg.LogRequestBody && len(requestBody) > 0 && int64(len(requestBody)) <= cfg.MaxBodySize {
 			redactedBody := redactSensitiveData(requestBody, cfg.SensitiveFields)
 			event.RequestBody = json.RawMessage(redactedBody)
 		}
